pyproject.toml
[project]
name = "academic-paper-api"
version = "0.1.0"
description = "Scrape academic papers by DOI and reconstruct them as Markdown"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "scrapling>=0.4",
    "pydoll-python>=2.20.2",
    "httpx>=0.27",
    "click>=8.0",
    "rich>=13.0",
]

[project.scripts]
paper-scrape = "academic_paper_api.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/academic_paper_api"]
